pyproject.toml
[project]
name = "aiogram-bot-template"
description = "Template for creating scalable bots with aiogram"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Forden", email = "24463229+Forden@users.noreply.github.com" }
]
maintainers = [
    { name = "Forden", email = "24463229+Forden@users.noreply.github.com" }
]
readme = "README.md"
packages = [{ include = "aiogram_bot_template" }]
dynamic = ["version", "dependencies"]

[project.urls]
homepage = "https://github.com/Forden/aiogram-bot-template"
repository = "https://github.com/Forden/aiogram-bot-template"

[project.scripts]
aiogram_bot_template = "aiogram_bot_template.bot:main"

[tool.poetry]
requires-poetry = ">=2.0"
package-mode = false
version = "1.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3"
environs = "^14"
redis = "^6"
asyncpg = "0.30.0"
pydantic = "^2"
structlog = "25.4.0"
aiojobs = "^1"
orjson = "^3.10"
tenacity = "^9"
pytz = "^2025"
aiohttp = "~3"
arrow = "^1"
apscheduler = "^3.11.2"
aiohttp-cors = "^0.8.1"
pyjwt = "^2.11.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
black = { extras = ["d"], version = "^25" }
ruff = "^0.12"
isort = "~6"
types-pytz = "^2025"
deptry = "0.23.0"
pip-audit = "^2"

[tool.black]
target-version = ["py310"]

[tool.isort]
profile = "black"
py_version = 310
force_alphabetical_sort_within_sections = true
group_by_package = true

[tool.mypy]
python_version = "3.10"
files = "bot.py"
plugins = "pydantic.mypy"
warn_unused_configs = true
ignore_missing_imports = false
follow_imports_for_stubs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_equality = true
strict_bytes = true
extra_checks = true
show_error_context = true
show_error_code_links = true
show_error_end = true
show_error_codes = true
pretty = true
namespace_packages = true
show_absolute_path = true

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
preview = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "E501", # line too long, controlled by black
    "ANN401", # attempt to replace mypy, ignored in favor of mypy
    "RET504", # opinionated
    "PLC0414", # breaks mypy unused imports
    "SIM108", # forces to use ternary operator
    "D", # unnecessary
    "I", # unnecessary
    "CPY001"# unnecessary
]
unfixable = ["B"]

[tool.ruff.format]
line-ending = "lf"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
